% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggEvalue.R
\name{ggEvalue}
\alias{ggEvalue}
\title{ggEvalue}
\usage{
ggEvalue(
  estimate,
  lo,
  hi,
  xlim = c(0, 7),
  ylim = c(0, 7),
  point.col = c("#90353b", "#1a476f"),
  text.col = c("#90353b", "#1a476f"),
  line.col = c("#0d680d", "#ff4602")
)
}
\arguments{
\item{estimate}{estimate value, for example, RR, OR and HR. class = numerical, length = 1.}

\item{lo}{lower 95\% CI of estimate value,  class = numerical, length = 1.}

\item{hi}{upper 95\% CI of estimate value,  class = numerical, length = 1.}

\item{xlim, ylim}{limitation for x axis with 2 numbers,
the first is the minimum value and the second is the maximum value. class = numerical, length = 2.}

\item{point.col, text.col, line.col}{colors for point, text and line. the first color is the CI and the second color is the E value.}
}
\value{
a ggplot object.
}
\description{
plot E value curve using ggplot2
}
